Fixed .gitignore
