--- .editorconfig
@@ -1,24 +1,0 @@
-root = true
-
-[*]
-charset = utf-8
-indent_style = space
-indent_size = 4
-insert_final_newline = true
-
-[*.java]
-trim_trailing_whitespace = true
-max_line_length = 120
-ij_java_generate_final_locals = true
-ij_java_generate_final_parameters = true
-
-[*.yml]
-indent_size = 2
-tab_width = 2
-
-[*.properties]
-ij_properties_keep_blank_lines = true
-
-[*.json]
-indent_size = 2
-tab_width = 2
