--- .gitattributes
@@ -1,8 +1,0 @@
-# Disable autocrlf on generated files, they always generate with LF
-# Add any extra files or paths here to make git stop saying they
-# are changed when only line endings change.
-src/generated/**/.cache/cache text eol=lf
-src/generated/**/*.json text eol=lf
-
-# Files provided to VM.
-src/main/scripts/**/*.lua text eol=lf
