--- .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,45 +1,0 @@
----
-name: Bug report
-about: Report an issue of the mod crashing or behaving in an unexpected way
-title: ''
-labels: bug
-assignees: ''
-
----
-
----
-name: Bug report
-about: Report an issue of the mod crashing or behaving in an unexpected way
-title: ''
-labels: bug
-assignees: ''
-
----
-
-**Describe the bug**
-A clear and concise description of what the bug is.
-
-**To Reproduce**
-Steps to reproduce the behavior:
-1. Singleplayer or multiplayer?
-2. Place block '....'
-3. Interact with '....'
-4. See error '...'
-
-**Log files**
-Please post your log files (client and server) to gist and link them here.
-If it's a crash report, please also upload the crash report to gist and link it here.
-
-**Expected behavior**
-A clear and concise description of what you expected to happen.
-
-**Screenshots**
-If applicable, add screenshots to help explain your problem.
-
-**Versions (please complete the following information):**
- - Minecraft: [e.g. 1.18.2]
- - Forge: [e.g. 40.0.40]
- - oc2r: [e.g. 0.1.7]
-
-**Additional context**
-Add any other context about the problem here.
