--- .github/workflows/build-1.19.2.yml
@@ -1,37 +1,0 @@
-name: Build
-on:
-  push:
-    branches:
-      - '1.19.2'
-  pull_request:
-    types: [ opened, synchronize, reopened ]
-jobs:
-  validate-gradle:
-    name: "Validate Gradle wrapper"
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          submodules: true # Clone with vs-core submodule
-      - uses: gradle/wrapper-validation-action@v1
-
-  test-server:
-    name: Test Server
-    strategy:
-      matrix:
-        serverType: [ 'forge' ]
-    runs-on: ubuntu-latest
-    timeout-minutes: 20 # Fail after 20 minutes
-    steps:
-      - name: Shallow Clone (--recurse-submodules)
-        uses: actions/checkout@v2
-        with:
-          submodules: recursive
-
-      - name: Setup Gradle & Caches
-        uses: "./.github/actions/setup-gradle"
-
-      - name: Test Server
-        uses: ValkyrienSkies/Minecraft-Architectury-Testing@v1.9
-        with:
-          serverType: ${{ matrix.serverType }}
