--- .github/workflows/build.yml
@@ -1,38 +1,0 @@
-#file: noinspection UndefinedAction,UndefinedParamsPresent
-name: build
-
-on:
-  workflow_dispatch:
-  push:
-    branches: [ '1.20.1' ]
-
-jobs:
-  build:
-    strategy:
-      matrix:
-        java: [ 17 ]
-        os: [ ubuntu-22.04 ]
-    runs-on: ${{ matrix.os }}
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: actions/cache@v4
-        with:
-          path: |
-            ~/.gradle/caches
-            ~/.gradle/wrapper
-          key: ${{ runner.os }}-gradle-${{ hashFiles('**/*.gradle*') }}
-          restore-keys: ${{ runner.os }}-gradle-
-
-      - name: Fix Windows
-        run: |
-          chmod +x gradlew
-      - name: build
-        run: |
-          ./gradlew build
-
-      - uses: actions/upload-artifact@v4
-        with:
-          name: oc2r.jar
-          path: build/libs/*-all.jar
