--- common/src/main/resources/oc2rvs.accesswidener
@@ -1,2 +1,0 @@
-accessWidener   v2  named
-
