--- forge/gradle.properties
@@ -1,3 +1,0 @@
-loom.platform=forge
-loader_platform=Forge
-kotlin.stdlib.default.dependency=false
