--- forge/src/main/resources/META-INF/mods.toml
@@ -1,49 +1,0 @@
-modLoader = "javafml"
-loaderVersion = "[3.12.0,)"
-issueTrackerURL = "https://github.com/ValkyrienSkies/Valkyrien-Skies-2/issues"
-license = "Apache-2.0"
-
-[[mods]]
-modId = "oc2rvs"
-version = "${version}"
-displayName = "OC2R: Valkyrien Skies 2"
-authors = ["lperkins2"]
-description = '''
-OC2R: Valkyrien Skies 2
-'''
-logoFile = "icon.png"
-
-[[dependencies.oc2rvs]]
-modId = "forge"
-mandatory = true
-versionRange = "[47.4.0,)"
-ordering = "NONE"
-side = "BOTH"
-
-[[dependencies.oc2rvs]]
-modId = "minecraft"
-mandatory = true
-versionRange = "[1.20.1,)"
-ordering = "NONE"
-side = "BOTH"
-
-[[dependencies.oc2rvs]]
-modId = "valkyrienskies"
-mandatory = true
-versionRange = "[${vs2_version},)"
-ordering = "AFTER"
-side = "BOTH"
-
-[[dependencies.oc2rvs]]
-modId = "oc2r"
-mandatory = true
-versionRange = "[2.2.4,)"
-ordering = "AFTER"
-side = "BOTH"
-
-[[dependencies.oc2rvs]]
-modId = "oc2rcore"
-mandatory = true
-versionRange = "[1.0.0,)"
-ordering = "AFTER"
-side = "BOTH"
