--- forge/src/main/resources/pack.mcmeta
@@ -1,6 +1,0 @@
-{
-  "pack": {
-    "description": "CC: VS",
-    "pack_format": 8
-  }
-}
