+++ gradle.properties
@@ -1,36 +1,53 @@
-org.gradle.jvmargs=-Xmx4096M
-# Identity
-mod_name=oc2rvs
-mod_id=oc2rvs
-mod_version=0.5.0
-enabled_platforms=forge
-archives_base_name=oc2rvs
-maven_group=com.therealm18studios
-mod_description=Support for OC2R in VS
-mod_license=Apache-2.0
-mod_author=github/therealm18
-sources=https://github.com/therealm18/oc2r-vs
-issue_tracker=https://github.com/therealm18/oc2r-vs/issues
-enabled_platforms=forge
-loader_platform=Common
-# Dependencies
-vs2_version=2.1.3-beta.1+a52f38bd68
-vs_core_version=1.1.0+b19b27c4a4
+org.gradle.jvmargs=-Xmx3G
+org.gradle.daemon=false
+# The Minecraft version must agree with the Forge version to get a valid artifact
 minecraft_version=1.20.1
-minecraft_sdk=Forge
-architectury_version=9.1.12
-fabric_loader_version=0.14.24
-fabric_api_version=0.90.7+1.20.1
-forge_version=1.20.1-47.2.0
-forge_kotlin_version=4.10.0
-kotlin_version=1.9.10
-cloth_config_version=6.4.90
-# Maven publishing
-vs_maven_url=
-vs_maven_username=
-vs_maven_password=
-# For release builds
-curse_api_key=
-curse_project_id=
-modrinth_api_key=
-modrinth_project_id=
+# The Minecraft version range can use any release version of Minecraft as bounds.
+# Snapshots, pre-releases, and release candidates are not guaranteed to sort properly
+# as they do not follow standard versioning conventions.
+minecraft_version_range=[1.20.1,1.21)
+# The Forge version must agree with the Minecraft version to get a valid artifact
+forge_version=47.4.8
+# The Forge version range can use any version of Forge as bounds or match the loader version range
+forge_version_range=[47,)
+# The loader version range can only use the major version of Forge/FML as bounds
+loader_version_range=[47,)
+# The mapping channel to use for mappings.
+# The default set of supported mapping channels are ["official", "snapshot", "snapshot_nodoc", "stable", "stable_nodoc"].
+# Additional mapping channels can be registered through the "channelProviders" extension in a Gradle plugin.
+#
+# | Channel   | Version              |                                                                                |
+# |-----------|----------------------|--------------------------------------------------------------------------------|
+# | official  | MCVersion            | Official field/method names from Mojang mapping files                          |
+# | parchment | YYYY.MM.DD-MCVersion | Open community-sourced parameter names and javadocs layered on top of official |
+#
+# You must be aware of the Mojang license when using the 'official' or 'parchment' mappings.
+# See more information here: https://github.com/MinecraftForge/MCPConfig/blob/master/Mojang.md
+#
+# Parchment is an unofficial project maintained by ParchmentMC, separate from Minecraft Forge.
+# Additional setup is needed to use their mappings, see https://parchmentmc.org/docs/getting-started
+mapping_channel=official
+# The mapping version to query from the mapping channel.
+# This must match the format required by the mapping channel.
+mapping_version=1.20.1
+# The unique mod identifier for the mod. Must be lowercase in English locale. Must fit the regex [a-z][a-z0-9_]{1,63}
+# Must match the String constant located in the main mod class annotated with @Mod.
+mod_id=oc2rvs
+# The human-readable display name for the mod.
+mod_name=OC2R: Valkyrien Skies 2
+# The license of the mod. Review your options at https://choosealicense.com/. All Rights Reserved is the default.
+mod_license=MIT
+# The mod version. See https://semver.org/
+mod_version=1.0.0
+# The group ID for the mod. It is only important when publishing as an artifact to a Maven repository.
+# This should match the base package used for the mod sources.
+# See https://maven.apache.org/guides/mini/guide-naming-conventions.html
+mod_group_id=com.therealm18studios.oc2rvs
+# The authors of the mod. This is a simple text string that is used for display purposes in the mod list.
+mod_authors=[lperkins2, therealm18, TechTastic]
+# The description of the mod. This is a simple multiline text string that is used for display purposes in the mod list.
+mod_description=Support for OC2R in VS
+
+vs2_version=2.3.0-beta.9
+vs_core_version=1.1.0+cf6990a85d
+kotlin_version=4.11.0
