--- minecraft.gradle
@@ -1,4 +1,0 @@
-ext {
-    minecraft_version = '1.20.1'
-    minecraft_sdk = 'forge'
-}
