+++ src/main/resources/data/oc2r/tags/items/device_needs_reboot.json
