+++ src/main/resources/data/oc2r/tags/items/devices/card.json
