+++ src/main/resources/oc2rvs.mixins.json
@@ -1,0 +1,17 @@
+{
+  "required": true,
+  "minVersion": "0.8",
+  "package": "com.therealm18studios.oc2rvs.mixin",
+  "compatibilityLevel": "JAVA_8",
+  "refmap": "oc2rvs.refmap.json",
+  "mixins": [
+  ],
+  "client": [
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  },
+  "overwrites": {
+    "requireAnnotations": true
+  }
+}
