+++ src/main/resources/pack.mcmeta
@@ -1,0 +1,6 @@
+{
+  "pack": {
+    "description": "OC2:R VS",
+    "pack_format": 15
+  }
+}
